# WMS_WebAPI/WMS_WebAPI/swagger/swaggerjson - Copy.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1"
  },
  "host": "localhost:50191",
  "schemes": [ "http", "https" ],
  "paths": {
...
    "/api/RateMaster/Rate_Insert": {
      "post": {
        "tags": [ "RateMaster" ],
        "operationId": "RateMaster_Rate_Insert",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_RateMaster" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
    "/api/Shifting/Shifting_insert": {
      "post": {
        "tags": [ "Shifting" ],
        "operationId": "Shifting_Shifting_insert",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_Shifting" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/Shifting/Shifting_List": {
      "post": {
        "tags": [ "Shifting" ],
        "operationId": "Shifting_Shifting_List",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_Shifting" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/Shifting/Shifting_Cancelled": {
      "post": {
        "tags": [ "Shifting" ],
        "operationId": "Shifting_Shifting_Cancelled",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_Shifting" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/Shifting/Rep_Shifting": {
      "post": {
        "tags": [ "Shifting" ],
        "operationId": "Shifting_Rep_Shifting",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_Shifting" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/Shifting/ShiftingEdit_List": {
      "post": {
        "tags": [ "Shifting" ],
        "operationId": "Shifting_ShiftingEdit_List",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_Shifting" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
    "/api/StorageAreaType/StorageAreaType_Insert_Update": {
      "post": {
        "tags": [ "StorageAreaType" ],
        "operationId": "StorageAreaType_StorageAreaType_Insert_Update",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_StorageAreaType" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
    "/api/TaxType/TaxType_Insert_Update": {
      "post": {
        "tags": [ "TaxType" ],
        "operationId": "TaxType_TaxType_Insert_Update",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Cls_TaxType" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
    "/api/TemperatureCategory/TemperatureCategories_Insert": {
      "post": {
        "tags": [ "TemperatureCategory" ],
        "operationId": "TemperatureCategory_TemperatureCategories_Insert",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_TemperatureCategories" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
    "/api/UserCompany/UserCompany_Insert": {
      "post": {
        "tags": [ "UserCompany" ],
        "operationId": "UserCompany_UserCompany_Insert",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_UserCompany" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/UserCompany/UserCompanyWareHouseList_select": {
      "post": {
        "tags": [ "UserCompany" ],
        "operationId": "UserCompany_UserCompanyWareHouseList_select",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_UserCompanyWareHouseList" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "cls_RateMaster": {
      "type": "object",
      "properties": {
        "RateID": {
          "format": "int32",
          "type": "integer"
        },
        "RateCode": { "type": "string" },
        "RateDescription": { "type": "string" },
        "StartDate": { "type": "string" },
        "EndDate": { "type": "string" },
        "CreatedBy": {
          "format": "int32",
          "type": "integer"
        },
        "LTD_RateList": {
          "type": "array",
          "items": { "$ref": "#/definitions/cls_TD_RateList" }
        }
      }
    },
    "cls_TD_RateList": {
      "type": "object",
      "properties": {
        "RateID": {
          "format": "int32",
          "type": "integer"
        },
        "ProductID": {
          "format": "int32",
          "type": "integer"
        },
        "ProductName": { "type": "string" },
        "ServiceID": {
          "format": "int32",
          "type": "integer"
        },
        "ServiceName": { "type": "string" },
        "Rate": {
          "format": "double",
          "type": "number"
        },
        "L_Rate": {
          "format": "double",
          "type": "number"
        },
        "TaxID": {
          "format": "int32",
          "type": "integer"
        },
        "BillingCycleID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "cls_Shifting": {
      "type": "object",
      "properties": {
        "remarks": { "type": "string" },
        "shiftingDID": {
          "format": "int32",
          "type": "integer"
        },
        "shiftingID": {
          "format": "int32",
          "type": "integer"
        },
        "warehouseID": {
          "format": "int32",
          "type": "integer"
        },
        "customerID": {
          "format": "int32",
          "type": "integer"
        },
        "createdBy": {
          "format": "int32",
          "type": "integer"
        },
        "shiftingDate": {
          "format": "date-time",
          "type": "string"
        },
        "loadingBy": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "cls_StorageAreaType": {
      "type": "object",
      "properties": {
        "StorageAreaTypeID": {
          "format": "int32",
          "type": "integer"
        },
        "StorageAreaType": { "type": "string" },
        "TemperatureCategoryID": {
          "format": "int32",
          "type": "integer"
        },
        "IsActive": { "type": "boolean" },
        "CreatedBy": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "Cls_TaxType": {
      "type": "object",
      "properties": {
        "TaxTypeID": {
          "format": "int32",
          "type": "integer"
        },
        "Code": { "type": "string" },
        "TaxTypeDescription": { "type": "string" },
        "isActive": { "type": "boolean" }
      }
    },
    "cls_TemperatureCategories": {
      "type": "object",
      "properties": {
        "TempCategoryID": {
          "format": "int32",
          "type": "integer"
        },
        "TempCategory": { "type": "string" },
        "MinTemp": {
          "format": "int32",
          "type": "integer"
        },
        "MaxTemp": {
          "format": "int32",
          "type": "integer"
        },
        "CreatedBy": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "cls_UserCompany": {
      "type": "object",
      "properties": {
        "UserCompanyID": {
          "format": "int32",
          "type": "integer"
        },
        "LTD_UserCompany": {
          "type": "array",
          "items": { "$ref": "#/definitions/cls_TD_UserCompany" }
        },
        "CreatedBy": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "cls_TD_UserCompany": {
      "type": "object",
      "properties": {
        "UserID": {
          "format": "int32",
          "type": "integer"
        },
        "CompanyID": {
          "format": "int32",
          "type": "integer"
        },
        "WarehouseID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "cls_UserCompanyWareHouseList": {
      "type": "object",
      "properties": {
        "UserID": {
          "format": "int32",
          "type": "integer"
        },
        "CompanyID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}